+++ kiplot/__main__.py
@@ -5,7 +5,7 @@ Usage:
   kiplot [-b BOARD] [-e SCHEMA] [-c CONFIG] [-d OUT_DIR] [-s PRE]
          [-q | -v...] [-i] [TARGET...]
   kiplot [-c PLOT_CONFIG] --list
-  kiplot [-b BOARD] [-d OUT_DIR] --example
+  kiplot [-b BOARD] [-d OUT_DIR] [-p] --example
   kiplot --help-list-outputs
   kiplot --help-output=HELP_OUTPUT
   kiplot --help-outputs
@@ -28,6 +28,7 @@ Options:
   --help-preflights                List supported preflights and details
   -i, --invert-sel                 Generate the outputs not listed as targets
   -l, --list                       List available outputs (in the config file)
+  -p, --copy-options               Copy plot options from the PCB file
   -q, --quiet                      Remove information logs
   -s PRE, --skip-pre PRE           Skip preflights, comma separated or `all`
   -v, --verbose                    Show debugging information
@@ -172,7 +173,10 @@ def main():
         sys.exit(0)
     if args.example:
         check_board_file(args.board_file)
-        create_example(args.board_file, GS.out_dir)
+        if args.copy_options and not args.board_file:
+            logger.error('Asked to copy options but no PCB specified.')
+            sys.exit(EXIT_BAD_ARGS)
+        create_example(args.board_file, GS.out_dir, args.copy_options)
         sys.exit(0)
 
     # Determine the YAML file