+++ kiplot/out_pdf.py
@@ -1,4 +1,4 @@
-from pcbnew import PLOT_FORMAT_PDF
+from pcbnew import (PLOT_FORMAT_PDF, FromMM, ToMM)
 from kiplot.out_any_layer import AnyLayer
 from kiplot.drill_marks import DrillMarks
 from kiplot.macros import macros, document, output_class  # noqa: F401
@@ -26,3 +26,17 @@ class PDF(AnyLayer, DrillMarks):
     def config(self, outdir, options, layers):
         AnyLayer.config(self, outdir, options, layers)
         DrillMarks.config(self)
+
+    def _configure_plot_ctrl(self, po, output_dir):
+        AnyLayer._configure_plot_ctrl(self, po, output_dir)
+        DrillMarks._configure_plot_ctrl(self, po, output_dir)
+        po.SetMirror(self.mirror_plot)
+        po.SetLineWidth(FromMM(self.line_width))
+        po.SetNegative(self.negative_plot)
+
+    def read_vals_from_po(self, po):
+        AnyLayer.read_vals_from_po(self, po)
+        DrillMarks.read_vals_from_po(self, po)
+        self.mirror_plot = po.GetMirror()
+        self.line_width = ToMM(po.GetLineWidth())
+        self.negative_plot = po.GetNegative()